requirements.txt
requests~=2.32.5
python-dotenv~=1.2.1
streamlit~=1.53.0
protobuf~=6.33.4
pandas~=2.3.3
plotly~=6.5.2